# perf(limit): refactor limit to use NarrowStateTransformation