--- a/internal/feature/flags.go
+++ b/internal/feature/flags.go
@@ -113,6 +113,18 @@ func OptimizeAggregateWindow() BoolFlag {
 	return optimizeAggregateWindow
 }
 
+var narrowTransformationLimit = feature.MakeBoolFlag(
+	"Narrow Transformation Limit",
+	"narrowTransformationLimit",
+	"Owen Nelson",
+	false,
+)
+
+// NarrowTransformationLimit - Enable the NarrowStateTransformation implementation of limit
+func NarrowTransformationLimit() BoolFlag {
+	return narrowTransformationLimit
+}
+
 // Inject will inject the Flagger into the context.
 func Inject(ctx context.Context, flagger Flagger) context.Context {
 	return feature.Inject(ctx, flagger)
@@ -127,6 +139,7 @@ var all = []Flag{
 	vectorizedMap,
 	narrowTransformationDifference,
 	optimizeAggregateWindow,
+	narrowTransformationLimit,
 }
 
 var byKey = map[string]Flag{
@@ -138,6 +151,7 @@ var byKey = map[string]Flag{
 	"vectorizedMap":                    vectorizedMap,
 	"narrowTransformationDifference":   narrowTransformationDifference,
 	"optimizeAggregateWindow":          optimizeAggregateWindow,
+	"narrowTransformationLimit":        narrowTransformationLimit,
 }
 
 // Flags returns all feature flags.