--- a/stdlib/contrib/bonitoo-io/tickscript/alert_test.flux
+++ b/stdlib/contrib/bonitoo-io/tickscript/alert_test.flux
@@ -68,5 +68,9 @@ tickscript_alert = (table=<-) =>
         )
         |> drop(columns: ["_time"])
 
-test _tickscript_alert = () =>
-    ({input: testing.loadStorage(csv: inData), want: testing.loadMem(csv: outData), fn: tickscript_alert})
+testcase tickscript_alert {
+    want = testing.loadMem(csv: outData)
+    got = testing.loadStorage(csv: inData) |> tickscript_alert()
+
+    testing.diff(want: want, got: got) |> yield()
+}