--- a/stdlib/contrib/bonitoo-io/tickscript/alert_with_topic_test.flux
+++ b/stdlib/contrib/bonitoo-io/tickscript/alert_with_topic_test.flux
@@ -69,5 +69,9 @@ tickscript_alert = (table=<-) =>
         )
         |> drop(columns: ["_time"])
 
-test _tickscript_alert = () =>
-    ({input: testing.loadStorage(csv: inData), want: testing.loadMem(csv: outData), fn: tickscript_alert})
+testcase tickscript_alert_with_topic {
+    want = testing.loadMem(csv: outData)
+    got = testing.loadStorage(csv: inData) |> tickscript_alert()
+
+    testing.diff(want: want, got: got) |> yield()
+}