--- a/stdlib/universe/limit_test.flux
+++ b/stdlib/universe/limit_test.flux
@@ -57,5 +57,6 @@ input = () =>
 testcase limit_1 {
     want = input()
     got = testing.loadMem(csv: outData)
+
     testing.diff(want: want, got: got) |> yield()
 }