Adds an (optimistically) optimized implementation for `limit` which leverages `NarrowStateTransformation` to read the incoming dataset as chunks.

The existing tests for limit were updated to run the established cases against both versions, the original impl _and the new Narrow_ version.

I also added "narrow" versions of the benchmarks in the `limit_test` package, though I'm not sure how strong signal these give. You can run these via `make test-bench`, then probably grep the output for `Limit`:

```
BenchmarkLimit_1N_1000-8                       1         133170217 ns/op        41040584 B/op     720007 allocs/op                                                                          
BenchmarkLimit_100N_1000-8                     1         137303909 ns/op        41040168 B/op     720006 allocs/op                                                                          
BenchmarkNarrowLimit_1N_1000-8                 1         105322194 ns/op        51309176 B/op     841101 allocs/op                                                                          
BenchmarkNarrowLimit_100N_1000-8               1         105539179 ns/op        51303848 B/op     841064 allocs/op  
```

Prior art: https://github.com/influxdata/flux/commit/fd7c66a1e9fd8cffcc9db9871222de03df11503f
Fixes: #3939

### Done checklist
- [ ] docs/SPEC.md updated **N/A**
- [x] Test cases written