@onelson 2022-03-01
After getting all the existing limit-related tests I could find running against both the original and new narrow implementations, I fixed one failure, but another remains:

```
--- FAIL: TestProcess_NarrowLimit_MultiBuffer (0.00s)
    limit_test.go:491: unexpected tables -want/+got
          []*executetest.Table{
                &{
                        ... // 2 identical fields
                        KeyValues: nil,
                        ColMeta:   {{Label: "_time", Type: s"time"}, {Label: "_value", Type: s"int"}},
                        Data: [][]interface{}{
                                {s"1970-01-01T00:00:00.000000020Z", int64(2)},
                                {s"1970-01-01T00:00:00.000000030Z", int64(3)},
        -                       {s"1970-01-01T00:00:00.000000040Z", int64(4)},
        +                       {s"1970-01-01T00:00:00.000000060Z", int64(6)},
        -                       {s"1970-01-01T00:00:00.000000050Z", int64(5)},
        +                       {s"1970-01-01T00:00:00.000000070Z", int64(7)},
                        },
                        Err:    nil,
                        Alloc:  nil,
                        IsDone: false,
                },
          }
```

I'm still looking into what's going wrong, but the test has _a lot_ of setup code to it. It'll take me a minute to get up to speed on what it's actually doing.